:tada: tag-variant migration setup
