--- @navikt/aksel/src/codemod/transforms/v8.0.0/color-update/tests/template.output.js
@@ -1,12 +1,0 @@
-import { Tag } from "@navikt/ds-react";
-
-const Component = () => {
-  return <div>
-    <Tag data-color="info" variant="outline">Text</Tag>
-    <Tag data-color="info" variant="moderate">Text</Tag>
-    <Tag data-color="info" variant="strong">Text</Tag>
-    <Tag data-color="info" variant="outline">Text</Tag>
-    <Tag data-color="info" variant="moderate">Text</Tag>
-    <Tag data-color="info" variant="strong">Text</Tag>
-  </div>;
-}
