+++ @navikt/aksel/src/codemod/transforms/v8.0.0/tag-variant/tests/all-variants.input.js
