+++ @navikt/aksel/src/codemod/transforms/v8.0.0/tag-variant/tests/index.test.ts
@@ -1,6 +1,6 @@
 import { check } from "../../../../utils/check";
 
-const migration = "color-update";
+const migration = "tag-variant";
 const fixtures = ["all-variants", "complex-cases", "ignored"];
 
 for (const fixture of fixtures) {
